+++ apps/restconf/restconf_native.h
@@ -155,6 +155,8 @@ typedef struct restconf_socket{
                                  * Set in restconf_callhome_cb
                                  */
     restconf_conn *rs_conns;  /* List of transient connect sockets */
+    char          *rs_from_addr; /* From IP address as seen by accept */
+
 } restconf_socket;
 
 /* Restconf handle